refactor : 쿠키 관련 피드백 반영, HomeController 분리
