+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,0 +1,22 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+@Controller
+public class AdminController {
+
+    @GetMapping("/admin/reservation")
+    public String adminReservation() {
+        return "admin/reservation";
+    }
+
+    @GetMapping("/admin/time")
+    public String time() {
+        return "admin/time";
+    }
+
+    @GetMapping("/admin/theme")
+    public String theme() {
+        return "admin/theme";
+    }
+}
