+++ src/main/java/roomescape/controller/HomeController.java
@@ -10,28 +10,4 @@ public class HomeController {
     public String home() {
         return "admin/index";
     }
-
-    @GetMapping("/admin/reservation")
-    public String adminReservation() {
-        return "admin/reservation";
-    }
-    @GetMapping("/admin/time")
-    public String time() {
-        return "admin/time";
-    }
-
-    @GetMapping("/admin/theme")
-    public String theme() {
-        return "admin/theme";
-    }
-
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "reservation";
-    }
-
-    @GetMapping("/login")
-    public String login() {
-        return "login";
-    }
 }
