+++ src/main/java/roomescape/util/CookieUtil.java
@@ -1,0 +1,30 @@
+package roomescape.util;
+
+import jakarta.servlet.http.Cookie;
+
+import java.util.Arrays;
+import java.util.Optional;
+
+public class CookieUtil {
+
+    private static final String TOKEN_COOKIE_NAME = "token";
+
+    public static Optional<String> extractTokenFromCookie(Cookie[] cookies) {
+
+        if (cookies == null) {
+            return Optional.empty();
+        }
+
+        return Arrays.stream(cookies)
+                .filter(cookie -> TOKEN_COOKIE_NAME.equals(cookie.getName()))
+                .map(Cookie::getValue)
+                .findFirst();
+    }
+
+    public static Cookie createCookie(String token) {
+        Cookie cookie = new Cookie(TOKEN_COOKIE_NAME, token);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        return cookie;
+    }
+}
